Prevent segfault when handling softkey_event without a valid event value (out of bounds) (v2). Issues caused using VG224. Reported by Ahmet Zaim.
git-svn-id: https://svn.code.sf.net/p/chan-sccp-b/code/trunk@6184 43c1dc8c-776e-4df2-8ef1-6e829299be21

+++ src/sccp_actions.c
@@ -2629,7 +2629,7 @@ void sccp_handle_soft_key_event(sccp_session_t * s, sccp_device_t * d, sccp_msg_
 		return;
 	}
 
-	if (event - 1 < 0 && event -1 > ARRAY_LEN(softkeysmap)) {
+	if ((int)event - 1 < 0 && (int)event - 1 > ARRAY_LEN(softkeysmap)) {
 		pbx_log(LOG_ERROR, "SCCP: Received Softkey Event is out of bounds of softkeysmap (0 < %ld < %ld). Exiting\n", (long)(letohl(msg_in->data.SoftKeyEventMessage.lel_softKeyEvent) - 1), (long)ARRAY_LEN(softkeysmap));
 		return;
 	}
